# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Students.API",
    "version": "1.0"
  },
    "paths": {
      "/users": {
        "put": {
            "tags": [
              "Users"
            ],
            "summary": "Update student.",
            "operationId": "UpdateStudent",
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "application/*+json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "boolean"
                    }
                  },
                  "application/json": {
                    "schema": {
                      "type": "boolean"
                    }
                  },
                  "text/json": {
                    "schema": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Create student.",
          "operationId": "CreateStudent",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUser"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUser"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }              
      },
      "/users/?": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get filtered student.",
          "operationId": "GetAllStudents",
          "parameters" : [ 
            {
              "name": "filter",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/users/{id}": {
        "delete": {
          "tags": [
            "Users"
          ],
          "summary": "Delete student with given ID",
          "operationId": "DeleteStudentById",
          "parameters" : [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "boolean"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get user by ID",
          "operationId": "GetStudentById",
          "parameters" : [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }       
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "uniqueItems": true,
              "nullable": false
            },
            "name": {
              "type": "string"
            },
            "course": {
                "type": "string"
            },
            "group": {
              "type": "string"
            }
          }
        },
        "CreateUser": {
          "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "course": {
                  "type": "string"
              },
              "group": {
                "type": "string"
              }
            }
          }
      }
    }    
}